requirements.txt
# FinMentor AI — Python Dependencies
# Install: pip install -r requirements.txt

# ── Core framework ────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1

# ── Anthropic LLM ────────────────────────────────────────────────────────────
anthropic==0.34.2
httpx==0.27.2

# ── Logging ───────────────────────────────────────────────────────────────────
loguru==0.7.2

# ── Numerical / ML ────────────────────────────────────────────────────────────
numpy

# ── RL (optional — only needed when rl_model.backend = "sb3") ─────────────────
# stable-baselines3==2.3.2
# torch==2.4.0                   # CPU build — swap for CUDA variant if GPU
# gymnasium==0.29.1

# ── ML fallback (optional — only needed when ml_fallback.backend = "xgboost") ─
# xgboost==2.1.1

# ── Redis memory store (optional — only needed when memory_backend = "redis") ──
# redis[asyncio]==5.0.8

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-mock==3.14.0
pytest-cov==5.0.0

# ── Development ────────────────────────────────────────────────────────────────
black==24.8.0
ruff==0.6.3
mypy==1.11.2
google-generativeai